Improve cpu performance, by getting rid of useless TerminalConsoleWriter
The `TerminalConsoleWriter` calls `LogQueues.getNextLogEvent("TerminalConsole")`.
The method `LogQueues.getNextLogEvent` is only used in the classes `MinecraftServerGUI` and `TerminalConsoleWriter`.
The method `LogQueus.getOrCreateQueue` is only used in `QueueLogAppender.createAppender`.
The Class `QueueLogAppender` is not used anywhere outside of itself.

So we have a thread that successfully does nothing! Great! Now why do we inherit this?
Because it's probably a leftover from spigot's Gui implementation, that might? need this.
We however removed that (or didn't include it). So we can remove this.